requirements/test.txt
-r base.txt

factory_boy == 2.9.2
flake8 == 3.4.1
pytest == 3.2.2
pytest-django == 3.1.2
